Update Editor.php
阻止同名覆盖

+++ Editor/Editor.php
@@ -16,9 +16,17 @@ class Editor {
 
 		if(isset($_POST['editorname']) && isset($_POST['editortext'])) {
 			$name = DIRECTORY_SEPARATOR == '\\' ? iconv('utf-8', 'gbk', $_POST['editorname']) : $_POST['editorname'];
-			$res = file_put_contents(CONTENTS_DIR.'/'.$name.'.md', $_POST['editortext']);
-			echo "<script>localStorage.text='';</script>";
-			if(!$res) die('文章保存失败');
+			$name = preg_replace('/[\/\\\<\>\*\?]/m','_',$name);
+			$file = CONTENTS_DIR.'/'.$name.'.md';
+			if(file_exists($file)){
+				echo "<meta http-equiv=\"content-type\" content=\"text/html; charset=utf-8\"><script>alert('已经有一篇名为 “".$name."” 的文章了，还请换个名字保存！');window.location.href='?edit';</script>";
+				die();
+			}else{
+				$res = file_put_contents(CONTENTS_DIR.'/'.$name.'.md', $_POST['editortext']);
+				echo "<script>localStorage.text='';</script>";
+				if(!$res) die('文章保存失败');
+			}
+			
 		}
 	}
 }
